Add Custom Theme to settings fixes #1471
