+++ src/components/ChatApp/MessageSection/MessageSection.react.js
@@ -9,12 +9,9 @@ import PropTypes from 'prop-types';
 import { addUrlProps, UrlQueryParamTypes } from 'react-url-query';
 import loadingGIF from '../../../images/loading.gif';
 import DialogSection from './DialogSection';
-import RaisedButton from 'material-ui/RaisedButton';
-import { CirclePicker } from 'react-color';
 import $ from 'jquery';
 import { Scrollbars } from 'react-custom-scrollbars';
 import TopBar from '../TopBar.react';
-import TextField from 'material-ui/TextField';
 import FloatingActionButton from 'material-ui/FloatingActionButton';
 import NavigateDown from 'material-ui/svg-icons/navigation/expand-more';
 import NavigateUp from 'material-ui/svg-icons/navigation/expand-less';
@@ -47,7 +44,6 @@ function getStateFromStores() {
     showLogin: false,
     openForgotPassword: false,
     showSignUp: false,
-    showThemeChanger: false,
     showHardwareChangeDialog: false,
     showHardware: false,
     showServerChangeDialog: false,
@@ -202,53 +198,6 @@ class MessageSection extends Component {
     };
   }
 
-  handleColorChange = (name, color) => {
-    // Current Changes
-  };
-  // Add Image as a background image
-  handleChangeBodyBackgroundImage = backImage => {
-    this.setState({ bodyBackgroundImage: backImage });
-  };
-
-  handleChangeMessageBackgroundImage = backImage => {
-    this.setState({ messageBackgroundImage: backImage });
-  };
-
-  // get the selected custom colour
-  handleChangeComplete = (name, color) => {
-    this.setState({ currTheme: 'custom' });
-    let currSettings = UserPreferencesStore.getPreferences();
-    let settingsChanged = {};
-    if (currSettings.Theme !== 'custom') {
-      settingsChanged.theme = 'custom';
-      Actions.settingsChanged(settingsChanged);
-    }
-    // Send these Settings to Server
-    let state = this.state;
-
-    if (name === 'header') {
-      state.header = color.hex;
-      this.customTheme.header = state.header.substring(1);
-    } else if (name === 'body') {
-      state.body = color.hex;
-      this.customTheme.body = state.body.substring(1);
-    } else if (name === 'pane') {
-      state.pane = color.hex;
-      this.customTheme.pane = state.pane.substring(1);
-    } else if (name === 'composer') {
-      state.composer = color.hex;
-      this.customTheme.composer = state.composer.substring(1);
-    } else if (name === 'textarea') {
-      state.textarea = color.hex;
-      this.customTheme.textarea = state.textarea.substring(1);
-    } else if (name === 'button') {
-      state.button = color.hex;
-      this.customTheme.button = state.button.substring(1);
-    }
-    this.setState(state);
-    document.body.style.setProperty('background-color', this.state.body);
-  };
-
   // Open Login Dialog
   handleOpen = () => {
     this.setState({
@@ -268,73 +217,12 @@ class MessageSection extends Component {
     this.child.closeOptions();
   };
 
-  handleRemoveUrlBody = () => {
-    if (!this.state.bodyBackgroundImage) {
-      this.setState({ SnackbarOpenBackground: true });
-      setTimeout(() => {
-        this.setState({
-          SnackbarOpenBackground: false,
-        });
-      }, 2500);
-    } else {
-      this.setState({
-        bodyBackgroundImage: '',
-      });
-    }
-  };
-
-  handleRemoveUrlMessage = () => {
-    if (!this.state.messageBackgroundImage) {
-      this.setState({ SnackbarOpenBackground: true });
-      setTimeout(() => {
-        this.setState({
-          SnackbarOpenBackground: false,
-        });
-      }, 2500);
-    } else {
-      this.setState({
-        messageBackgroundImage: '',
-      });
-    }
-  };
-
-  handleRemoveUrlBody = () => {
-    if (!this.state.bodyBackgroundImage) {
-      this.setState({ SnackbarOpenBackground: true });
-      setTimeout(() => {
-        this.setState({
-          SnackbarOpenBackground: false,
-        });
-      }, 2500);
-    } else {
-      this.setState({
-        bodyBackgroundImage: '',
-      });
-      this.handleChangeBodyBackgroundImage('');
-    }
-  };
-
-  handleRemoveUrlMessage = () => {
-    if (!this.state.messageBackgroundImage) {
-      this.setState({ SnackbarOpenBackground: true });
-      setTimeout(() => {
-        this.setState({
-          SnackbarOpenBackground: false,
-        });
-      }, 2500);
-    } else {
-      this.setState({
-        messageBackgroundImage: '',
-      });
-    }
-  };
   // Close all dialog boxes
   handleClose = () => {
     var prevThemeSettings = this.state.prevThemeSettings;
     this.setState({
       showLogin: false,
       showSignUp: false,
-      showThemeChanger: false,
       openForgotPassword: false,
     });
 
@@ -386,7 +274,6 @@ class MessageSection extends Component {
       this.setState({
         showLogin: false,
         showSignUp: false,
-        showThemeChanger: false,
         openForgotPassword: false,
       });
     }
@@ -395,137 +282,11 @@ class MessageSection extends Component {
     this.setState({
       showLogin: false,
       showSignUp: false,
-      showThemeChanger: false,
       openForgotPassword: false,
       tour: false,
     });
     cookies.set('visited', true, { path: '/' });
   };
-  // Save Custom Theme settings on server
-  saveThemeSettings = () => {
-    let customData = '';
-    Object.keys(this.customTheme).forEach(key => {
-      customData = customData + this.customTheme[key] + ',';
-    });
-
-    let settingsChanged = {};
-    settingsChanged.theme = 'custom';
-    settingsChanged.customThemeValue = customData;
-    if (this.state.bodyBackgroundImage || this.state.messageBackgroundImage) {
-      settingsChanged.backgroundImage =
-        this.state.bodyBackgroundImage +
-        ',' +
-        this.state.messageBackgroundImage;
-    }
-    Actions.settingsChanged(settingsChanged);
-    this.setState({ currTheme: 'custom' });
-    this.setState({
-      showLogin: false,
-      showSignUp: false,
-      showThemeChanger: false,
-      openForgotPassword: false,
-    });
-  };
-
-  handleRestoreDefaultThemeClick = () => {
-    this.setState({
-      showLogin: false,
-      showSignUp: false,
-      showThemeChanger: false,
-      openForgotPassword: false,
-    });
-    var prevTheme = this.state.prevThemeSettings.currTheme;
-    var currTheme = this.state.currTheme;
-    if (
-      (currTheme === 'custom' && prevTheme === 'dark') ||
-      currTheme === 'dark'
-    ) {
-      this.applyDarkTheme();
-    } else {
-      this.applyLightTheme();
-    }
-  };
-
-  applyLightTheme = () => {
-    this.setState({
-      prevThemeSettings: null,
-      body: '#fff',
-      header: '#4285f4',
-      composer: '#f3f2f4',
-      pane: '#f3f2f4',
-      textarea: '#fff',
-      button: '#4285f4',
-      currTheme: 'light',
-    });
-    let customData = '';
-    Object.keys(this.customTheme).forEach(key => {
-      customData = customData + this.customTheme[key] + ',';
-    });
-
-    let settingsChanged = {};
-    settingsChanged.theme = 'light';
-    settingsChanged.customThemeValue = customData;
-    if (this.state.bodyBackgroundImage || this.state.messageBackgroundImage) {
-      settingsChanged.backgroundImage =
-        this.state.bodyBackgroundImage +
-        ',' +
-        this.state.messageBackgroundImage;
-    }
-    Actions.settingsChanged(settingsChanged);
-  };
-
-  applyDarkTheme = () => {
-    this.setState({
-      prevThemeSettings: null,
-      body: '#fff',
-      header: '#4285f4',
-      composer: '#f3f2f4',
-      pane: '#f3f2f4',
-      textarea: '#fff',
-      button: '#4285f4',
-      currTheme: 'dark',
-    });
-    let customData = '';
-    Object.keys(this.customTheme).forEach(key => {
-      customData = customData + this.customTheme[key] + ',';
-    });
-
-    let settingsChanged = {};
-    settingsChanged.theme = 'dark';
-    settingsChanged.customThemeValue = customData;
-    if (this.state.bodyBackgroundImage || this.state.messageBackgroundImage) {
-      settingsChanged.backgroundImage =
-        this.state.bodyBackgroundImage +
-        ',' +
-        this.state.messageBackgroundImage;
-    }
-    Actions.settingsChanged(settingsChanged);
-  };
-
-  handleThemeChanger = () => {
-    this.setState({ showThemeChanger: true });
-    // save the previous theme settings
-    if (this.state.currTheme === 'light') {
-      // remove the previous custom theme memory
-      this.applyLightTheme();
-    }
-    var prevThemeSettings = {};
-    var state = this.state;
-    prevThemeSettings.currTheme = state.currTheme;
-    if (state.currTheme === 'custom') {
-      prevThemeSettings.bodyColor = state.body;
-      prevThemeSettings.TopBarColor = state.header;
-      prevThemeSettings.composerColor = state.composer;
-      prevThemeSettings.messagePane = state.pane;
-      prevThemeSettings.textArea = state.textarea;
-      prevThemeSettings.buttonColor = state.button;
-      prevThemeSettings.bodyBackgroundImage = state.bodyBackgroundImage;
-      prevThemeSettings.messageBackgroundImage = state.messageBackgroundImage;
-    }
-    this.setState({ prevThemeSettings });
-    this.child.closeOptions();
-  };
-
   // Show forgot password dialog
   forgotPasswordChanged = () => {
     this.setState({
@@ -751,6 +512,7 @@ class MessageSection extends Component {
     var composerColor;
     var messagePane;
     var textArea;
+    // eslint-disable-next-line
     var buttonColor;
     var textColor;
 
@@ -840,141 +602,6 @@ class MessageSection extends Component {
       }
     }
 
-    const actions = (
-      <RaisedButton
-        label={<Translate text="Cancel" />}
-        backgroundColor={
-          UserPreferencesStore.getTheme() === 'light' ? '#4285f4' : '#19314B'
-        }
-        labelColor="#fff"
-        width="200px"
-        keyboardFocused={true}
-        onTouchTap={this.handleClose}
-      />
-    );
-
-    const customSettingsDone = (
-      <div>
-        <RaisedButton
-          label={<Translate text="Save" />}
-          backgroundColor={buttonColor ? buttonColor : '#4285f4'}
-          labelColor="#fff"
-          width="200px"
-          keyboardFocused={false}
-          onTouchTap={this.saveThemeSettings}
-          style={{ margin: '0 5px' }}
-        />
-        <RaisedButton
-          label={<Translate text="Reset" />}
-          backgroundColor={buttonColor ? buttonColor : '#4285f4'}
-          labelColor="#fff"
-          width="200px"
-          keyboardFocused={false}
-          onTouchTap={this.handleRestoreDefaultThemeClick}
-          style={{ margin: '0 5px' }}
-        />
-      </div>
-    );
-    // Custom Theme feature Component
-    const componentsList = [
-      { id: 1, component: 'header', name: 'Header' },
-      { id: 2, component: 'pane', name: 'Message Pane' },
-      { id: 3, component: 'body', name: 'Body' },
-      { id: 4, component: 'composer', name: 'Composer' },
-      { id: 5, component: 'textarea', name: 'Textarea' },
-      { id: 6, component: 'button', name: 'Button' },
-    ];
-
-    const components = componentsList.map(component => {
-      return (
-        <div key={component.id} className="circleChoose">
-          <h4>
-            <Translate text="Color of" /> <Translate text={component.name} />:
-          </h4>
-          <CirclePicker
-            color={component}
-            width={'100%'}
-            colors={[
-              '#f44336',
-              '#e91e63',
-              '#9c27b0',
-              '#673ab7',
-              '#3f51b5',
-              '#2196f3',
-              '#03a9f4',
-              '#00bcd4',
-              '#009688',
-              '#4caf50',
-              '#8bc34a',
-              '#cddc39',
-              '#ffeb3b',
-              '#ffc107',
-              '#ff9800',
-              '#ff5722',
-              '#795548',
-              '#607d8b',
-              '#0f0f0f',
-              '#ffffff',
-            ]}
-            onChangeComplete={this.handleChangeComplete.bind(
-              this,
-              component.component,
-            )}
-            onChange={this.handleColorChange.bind(this, component.id)}
-          />
-
-          <TextField
-            name="backgroundImg"
-            style={{
-              display: component.component === 'body' ? 'block' : 'none',
-            }}
-            onChange={(e, value) => this.handleChangeBodyBackgroundImage(value)}
-            value={this.state.bodyBackgroundImage}
-            floatingLabelText={<Translate text="Body Background Image URL" />}
-          />
-          <RaisedButton
-            name="removeBackgroundBody"
-            key={'RemoveBody'}
-            label={<Translate text="Remove URL" />}
-            style={{
-              display: component.component === 'body' ? 'block' : 'none',
-              width: '150px',
-            }}
-            backgroundColor={buttonColor ? buttonColor : '#4285f4'}
-            labelColor="#fff"
-            keyboardFocused={true}
-            onTouchTap={this.handleRemoveUrlBody}
-          />
-          <TextField
-            name="messageImg"
-            style={{
-              display: component.component === 'pane' ? 'block' : 'none',
-            }}
-            onChange={(e, value) =>
-              this.handleChangeMessageBackgroundImage(value)
-            }
-            value={this.state.messageBackgroundImage}
-            floatingLabelText={
-              <Translate text="Message Background Image URL" />
-            }
-          />
-          <RaisedButton
-            name="removeBackgroundMessage"
-            key={'RemoveMessage'}
-            label={<Translate text="Remove URL" />}
-            style={{
-              display: component.component === 'pane' ? 'block' : 'none',
-              width: '150px',
-            }}
-            backgroundColor={buttonColor ? buttonColor : '#4285f4'}
-            labelColor="#fff"
-            keyboardFocused={true}
-            onTouchTap={this.handleRemoveUrlMessage}
-          />
-        </div>
-      );
-    });
-
     let speechOutput = UserPreferencesStore.getSpeechOutput();
     let speechOutputAlways = UserPreferencesStore.getSpeechOutputAlways();
 
@@ -1011,7 +638,6 @@ class MessageSection extends Component {
               ref={instance => {
                 this.child = instance;
               }}
-              handleThemeChanger={this.handleThemeChanger}
               handleOpen={this.handleOpen}
               handleSignUp={this.handleSignUp}
               handleOptions={this.handleOptions}
@@ -1101,12 +727,8 @@ class MessageSection extends Component {
                 openLogin={this.state.showLogin}
                 openSignUp={this.state.showSignUp}
                 openForgotPassword={this.state.openForgotPassword}
-                openThemeChanger={this.state.showThemeChanger}
-                ThemeChangerComponents={components}
                 bodyStyle={bodyStyle}
-                actions={actions}
                 handleSignUp={this.handleSignUp}
-                customSettingsDone={customSettingsDone}
                 onRequestClose={() => this.handleClose}
                 onRequestCloseTour={() => this.handleCloseTour}
                 onSaveThemeSettings={() => this.handleSaveTheme}
